added an XMLReader implementation
lots of wavering around instances and encapsulation,
but I've decided to go with the generally-accepted best-practice
of isolating the components and binding them
in stateless method calls only, for now.

naming is terrible, and next i'll try to fix it.
the rub is trying to find three different names
for XMLReader and Parser, while still being consistent to the legacy.