SchemaUtil and test for XBRL schemas
failing currently; i'll fix the schema's references now.
I believe it's mostly http v.s. https;
i'm sure i could fix it by having the parser follow redirects,
but this is an exercise to build an excel-compatible schema,
which appears also not to follow redirects
(and presumably harder to configure)